handling windows-1251 encoding

+++ reformat-csv.rb
@@ -1,14 +1,16 @@
 #!/usr/bin/env ruby
 
-ARGV.each do |a|
-  $infile = "#{a}.txt"
-  $outfile = "#{a}.csv"
-end
+$infile = ARGV[0]
+$outfile = "#{$infile}"
 
-File.open( $outfile, 'w' ) do |f| 
-  File.foreach( $infile ) do |line|
-    line1 = line.chomp.gsub(/\"/, '""')
-    line2 = line1.gsub(/\|/, '"|"')
-    f.puts( '"'+line2+'"'  )
+if $infile && File.exist?($infile)
+  $outfile = "new-#{$infile}"
+  File.open( $outfile, 'w' ) do |f| 
+    File.foreach( $infile, "r:windows-1251:utf-8" ) do |line|
+      p line
+      line1 = line.chomp.gsub(/\"/, '""')
+      line2 = line1.gsub(/\|/, '"|"')
+      f.puts( '"'+line2+'"'  )
+    end
   end
 end
